Struct capability::WithCap

pub struct WithCap<T> { /* private fields */ }

A wrapper that holds a type with a capability.

Implementations

impl<T> WithCap<T>

pub fn new(inner: T, cap: Cap) -> Self

Create a new instance with the given capability.

pub const fn cap(&self) -> Cap

Get the capability.

pub const fn can_access(&self, cap: Cap) -> bool

Check if the inner data can be accessed with the given capability.

Examples

use capability::{Cap, WithCap};

let data = WithCap::new(42, Cap::READ);

assert!(data.can_access(Cap::READ));
assert!(!data.can_access(Cap::WRITE));

pub unsafe fn access_unchecked(&self) -> &T

Access the inner value without capability check.

Safety

Caller must ensure not to violate the capability.

pub const fn access(&self, cap: Cap) -> Result<&T, CapError>

Access the inner value with the given capability, or return CapError if cannot access.

Examples

use capability::{Cap, CapError, WithCap};

let data = WithCap::new(42, Cap::READ);

assert_eq!(data.access(Cap::READ).unwrap(), &42);
assert_eq!(data.access(Cap::WRITE).err(), Some(CapError::default()));

pub fn access_or_err<E>(&self, cap: Cap, err: E) -> Result<&T, E>

Access the inner value with the given capability, or return the given err if cannot access.

Examples

use capability::{Cap, WithCap};

let data = WithCap::new(42, Cap::READ);

assert_eq!(data.access_or_err(Cap::READ, "cannot read").unwrap(), &42);
assert_eq!(data.access_or_err(Cap::WRITE, "cannot write").err(), Some("cannot write"));